Split DCAT forms and views to directories
