+++ tests/dcat/forms/test_dataset_forms.py
@@ -12,7 +12,7 @@ from vitrina.classifiers.models import (
 from vitrina.datasets.form_helpers import DATASET_STANDARD_URI
 from vitrina.datasets.factories import ContactFactory, DatasetFactory
 from vitrina.datasets.models import Dataset
-from vitrina.dcat.forms import (
+from vitrina.dcat.forms.dataset_forms import (
     DatasetResourceForm,
     InformationSystemResourceForm,
     ServiceResourceForm,
